+++ Calculadora/src/ICalculadora.java
@@ -2,6 +2,6 @@ public interface ICalculadora {
     int sumar(int a, int b);
     int restar(int a, int b);
     int multiplicar(int a, int b);
-    int dividir(int a, int b) throws Exception;
+    double dividir(int a, int b) throws Exception;
 }
